requirements.txt
blinker==1.9.0
click==8.3.0
colorzero==2.0
Flask==3.1.2
gpiozero==2.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.3
paho-mqtt==2.1.0
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
RPi.GPIO==0.7.1
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
